centralize all reading/writing
